requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
pyyaml>=6.0
transformers>=4.30.0
torch>=2.0.0
jinja2>=3.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
fpdf2>=2.7.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
colorama>=0.4.0
tqdm>=4.65.0